implemented new method and tests
