+++ test/Backstage_Passes/qualityTwo.test.js
@@ -1,4 +1,4 @@
-const { Shop, Item } = require("../../src/gilded_rose3");
+const { Shop, Item } = require("../../src/gilded_rose");
 
 describe("Qualidade aumenta em 2", function () {
   it("Quando data venda <= 10", function () {
